Migrate to C++20 and v142 (#709)
* Fit for SDKv143

* The big migration pt. 1

* Fix (?) nightly build workflow

* Friendship ended with SFINAE, concepts are Phobos' new best friend

* Disable senseless complaints about uninited members

should somehow only disable it for YRpp but oh well

* Fix concept

* Automatically enable stream debugging for debug builds

* Clean up and localize code

* Delete Hooks.Savegame.cpp from vcxproj

* Update c_cpp_properties

* Fix compilation and switch to v142

* Adjust .vxconfig

* Adjust readme

* Adjust action version

* screw off EditorConfig

* Update Dir with STL header <bit> in C++20

* Fix deprecated macro usage

* Add Utilities/Coroutine.h

Update YRPP

* Update .vsconfig

---------

+++ src/Ext/BuildingType/Body.cpp
@@ -53,7 +53,7 @@ int BuildingTypeExt::GetEnhancedPower(BuildingClass* pBuilding, HouseClass* pHou
 	{
 		if (pExt->PowerPlantEnhancer_Buildings.Contains(pBuilding->Type))
 		{
-			fFactor *= std::pow(pExt->PowerPlantEnhancer_Factor.Get(1.0f), nCount);
+			fFactor *= std::powf(pExt->PowerPlantEnhancer_Factor.Get(1.0f), static_cast<float>(nCount));
 			nAmount += pExt->PowerPlantEnhancer_Amount.Get(0) * nCount;
 		}
 	}
